HP-UX fixes
git-svn-id: https://svn.wxwidgets.org/svn/wx/wxWidgets/trunk@3995 c3d73ce0-8a6f-49c7-b76d-6d57e0e08775

+++ src/generic/dcpsg.cpp
@@ -476,7 +476,7 @@ void wxPostScriptDC::DoDrawArc (long x1, long y1, long x2, long y2, long xc, lon
 
     long dx = x1 - xc;
     long dy = y1 - yc;
-    long radius = (long) sqrt(dx*dx+dy*dy);
+    long radius = (long) sqrt( (double)(dx*dx+dy*dy) );
     double alpha1, alpha2;
 
     if (x1 == x2 && y1 == y2)
